+++ CHANGELOG.md
@@ -8,6 +8,18 @@
 ### Changed
 ### Fixed
 ```
+
+## [Unrealased] - yyyy-mm-dd
+
+### Added
+
+- Add time command
+
+### Changed
+
+- Remove action alias
+- Change copy action to update
+
 ## [1.43.3] - 2021-07-29
 
 ### Fixed